clean the depot add poetry env add a makefile add a readme create folder/organise
