+++ Deplacement/AVANCER_G_D_sans_CAM_DEMO.py
@@ -1,8 +1,7 @@
 import os
 
 
-
-if os.name == 'nt':
+if os.name == "nt":
 
     import msvcrt
 
@@ -33,15 +32,12 @@ else:
         return ch
 
 
-
 from dynamixel_sdk import *  # Uses Dynamixel SDK library
-
 
 
 # Define Dynamixel model
 
-MY_DXL = 'X_SERIES'  # X330 (5.0 V recommended), X430, X540, 2X430
-
+MY_DXL = "X_SERIES"  # X330 (5.0 V recommended), X430, X540, 2X430
 
 
 # Control table address
@@ -59,7 +55,6 @@ LEN_PRESENT_POSITION = 4  # Data Byte Length
 ADDR_PROFILE_VELOCITY = 112  # Profile Velocity
 
 ADDR_PROFILE_ACCELERATION = 108  # Profile Acceleration
-
 
 
 # Movement positions
@@ -93,7 +88,6 @@ DXL_MINIMUM_POSITION_VALUE_GD11 = 3000
 DXL_MAXIMUM_POSITION_VALUE_GD11 = 2500
 
 
-
 # Set the profile velocity and acceleration
 
 PROFILE_VELOCITY = 70
@@ -101,15 +95,12 @@ PROFILE_VELOCITY = 70
 PROFILE_ACCELERATION = 30
 
 
-
 BAUDRATE = 1000000
-
 
 
 # Protocol version
 
 PROTOCOL_VERSION = 2.0
-
 
 
 # Dynamixel IDs for lift and rotate
@@ -119,11 +110,9 @@ LIFT_SERVOS = [2, 4, 6, 8, 10, 12]
 ROTATE_SERVOS = [id - 1 for id in LIFT_SERVOS]
 
 
-
 # Port
 
-DEVICENAME = '/dev/ttyUSB0'
-
+DEVICENAME = "/dev/ttyUSB0"
 
 
 TORQUE_ENABLE = 1
@@ -133,43 +122,26 @@ TORQUE_DISABLE = 0
 DXL_MOVING_STATUS_THRESHOLD = 75
 
 
-
 # Goal positions for lifting and rotating
 
 lift_positions = {
-
     2: [DXL_MINIMUM_POSITION_VALUE_HB, DXL_MAXIMUM_POSITION_VALUE_HB],
-
     4: [DXL_MINIMUM_POSITION_VALUE_HB, DXL_MAXIMUM_POSITION_VALUE_HB],
-
     6: [DXL_MINIMUM_POSITION_VALUE_HB, DXL_MAXIMUM_POSITION_VALUE_HB],
-
     8: [DXL_MINIMUM_POSITION_VALUE_HB, DXL_MAXIMUM_POSITION_VALUE_HB],
-
     10: [DXL_MINIMUM_POSITION_VALUE_HB, DXL_MAXIMUM_POSITION_VALUE_HB],
-
-    12: [DXL_MINIMUM_POSITION_VALUE_HB, DXL_MAXIMUM_POSITION_VALUE_HB]
-
+    12: [DXL_MINIMUM_POSITION_VALUE_HB, DXL_MAXIMUM_POSITION_VALUE_HB],
 }
-
 
 
 rotate_positions = {
-
     1: [DXL_MINIMUM_POSITION_VALUE_GD1, DXL_MAXIMUM_POSITION_VALUE_GD1],
-
     3: [DXL_MINIMUM_POSITION_VALUE_GD3, DXL_MAXIMUM_POSITION_VALUE_GD3],
-
     5: [DXL_MINIMUM_POSITION_VALUE_GD5, DXL_MAXIMUM_POSITION_VALUE_GD5],
-
     7: [DXL_MINIMUM_POSITION_VALUE_GD7, DXL_MAXIMUM_POSITION_VALUE_GD7],
-
     9: [DXL_MINIMUM_POSITION_VALUE_GD9, DXL_MAXIMUM_POSITION_VALUE_GD9],
-
-    11: [DXL_MINIMUM_POSITION_VALUE_GD11, DXL_MAXIMUM_POSITION_VALUE_GD11]
-
+    11: [DXL_MINIMUM_POSITION_VALUE_GD11, DXL_MAXIMUM_POSITION_VALUE_GD11],
 }
-
 
 
 # Initialize PortHandler instance
@@ -177,23 +149,23 @@ rotate_positions = {
 portHandler = PortHandler(DEVICENAME)
 
 
-
 # Initialize PacketHandler instance
 
 packetHandler = PacketHandler(PROTOCOL_VERSION)
 
 
-
 # Initialize GroupSyncWrite instance
 
-groupSyncWrite = GroupSyncWrite(portHandler, packetHandler, ADDR_GOAL_POSITION, LEN_GOAL_POSITION)
-
+groupSyncWrite = GroupSyncWrite(
+    portHandler, packetHandler, ADDR_GOAL_POSITION, LEN_GOAL_POSITION
+)
 
 
 # Initialize GroupSyncRead instance for Present Position
 
-groupSyncRead = GroupSyncRead(portHandler, packetHandler, ADDR_PRESENT_POSITION, LEN_PRESENT_POSITION)
-
+groupSyncRead = GroupSyncRead(
+    portHandler, packetHandler, ADDR_PRESENT_POSITION, LEN_PRESENT_POSITION
+)
 
 
 # Open port
@@ -209,7 +181,6 @@ if not portHandler.openPort():
 print("Succeeded to open the port")
 
 
-
 # Set port baudrate
 
 if not portHandler.setBaudRate(BAUDRATE):
@@ -223,53 +194,66 @@ if not portHandler.setBaudRate(BAUDRATE):
 print("Succeeded to change the baudrate")
 
 
-
 # Enable Dynamixel Torque
 
 for dxl_id in LIFT_SERVOS + ROTATE_SERVOS:
 
-    dxl_comm_result, dxl_error = packetHandler.write1ByteTxRx(portHandler, dxl_id, ADDR_TORQUE_ENABLE, TORQUE_ENABLE)
+    dxl_comm_result, dxl_error = packetHandler.write1ByteTxRx(
+        portHandler, dxl_id, ADDR_TORQUE_ENABLE, TORQUE_ENABLE
+    )
 
     if dxl_comm_result != COMM_SUCCESS:
 
-        print(f"Error enabling torque for Dynamixel {dxl_id}: {packetHandler.getTxRxResult(dxl_comm_result)}")
+        print(
+            f"Error enabling torque for Dynamixel {dxl_id}: {packetHandler.getTxRxResult(dxl_comm_result)}"
+        )
 
     elif dxl_error != 0:
 
-        print(f"Error enabling torque for Dynamixel {dxl_id}: {packetHandler.getRxPacketError(dxl_error)}")
+        print(
+            f"Error enabling torque for Dynamixel {dxl_id}: {packetHandler.getRxPacketError(dxl_error)}"
+        )
 
     else:
 
         print(f"Dynamixel#{dxl_id} has been successfully connected")
 
 
-
 # Set profile velocity and acceleration
 
 for dxl_id in LIFT_SERVOS + ROTATE_SERVOS:
 
-    dxl_comm_result, dxl_error = packetHandler.write4ByteTxRx(portHandler, dxl_id, ADDR_PROFILE_VELOCITY, PROFILE_VELOCITY)
+    dxl_comm_result, dxl_error = packetHandler.write4ByteTxRx(
+        portHandler, dxl_id, ADDR_PROFILE_VELOCITY, PROFILE_VELOCITY
+    )
 
     if dxl_comm_result != COMM_SUCCESS:
 
-        print(f"Error setting profile velocity for Dynamixel {dxl_id}: {packetHandler.getTxRxResult(dxl_comm_result)}")
+        print(
+            f"Error setting profile velocity for Dynamixel {dxl_id}: {packetHandler.getTxRxResult(dxl_comm_result)}"
+        )
 
     elif dxl_error != 0:
 
-        print(f"Error setting profile velocity for Dynamixel {dxl_id}: {packetHandler.getRxPacketError(dxl_error)}")
+        print(
+            f"Error setting profile velocity for Dynamixel {dxl_id}: {packetHandler.getRxPacketError(dxl_error)}"
+        )
 
-
-
-    dxl_comm_result, dxl_error = packetHandler.write4ByteTxRx(portHandler, dxl_id, ADDR_PROFILE_ACCELERATION, PROFILE_ACCELERATION)
+    dxl_comm_result, dxl_error = packetHandler.write4ByteTxRx(
+        portHandler, dxl_id, ADDR_PROFILE_ACCELERATION, PROFILE_ACCELERATION
+    )
 
     if dxl_comm_result != COMM_SUCCESS:
 
-        print(f"Error setting profile acceleration for Dynamixel {dxl_id}: {packetHandler.getTxRxResult(dxl_comm_result)}")
+        print(
+            f"Error setting profile acceleration for Dynamixel {dxl_id}: {packetHandler.getTxRxResult(dxl_comm_result)}"
+        )
 
     elif dxl_error != 0:
 
-        print(f"Error setting profile acceleration for Dynamixel {dxl_id}: {packetHandler.getRxPacketError(dxl_error)}")
-
+        print(
+            f"Error setting profile acceleration for Dynamixel {dxl_id}: {packetHandler.getRxPacketError(dxl_error)}"
+        )
 
 
 # Add parameter storage for present position values
@@ -283,7 +267,6 @@ for dxl_id in LIFT_SERVOS + ROTATE_SERVOS:
         print(f"[ID:{dxl_id:03d}] groupSyncRead addparam failed")
 
         quit()
-
 
 
 def move_to_position(ids_positions):
@@ -302,18 +285,11 @@ def move_to_position(ids_positions):
 
             continue
 
-
-
         param_goal_position = [
-
             DXL_LOBYTE(DXL_LOWORD(goal_position)),
-
             DXL_HIBYTE(DXL_LOWORD(goal_position)),
-
             DXL_LOBYTE(DXL_HIWORD(goal_position)),
-
-            DXL_HIBYTE(DXL_HIWORD(goal_position))
-
+            DXL_HIBYTE(DXL_HIWORD(goal_position)),
         ]
 
         dxl_addparam_result = groupSyncWrite.addParam(dxl_id, param_goal_position)
@@ -324,8 +300,6 @@ def move_to_position(ids_positions):
 
             quit()
 
-
-
     # Syncwrite goal position
 
     dxl_comm_result = groupSyncWrite.txPacket()
@@ -334,13 +308,9 @@ def move_to_position(ids_positions):
 
         print(f"Error during SyncWrite: {packetHandler.getTxRxResult(dxl_comm_result)}")
 
-
-
     # Clear syncwrite parameter storage
 
     groupSyncWrite.clearParam()
-
-
 
     while True:
 
@@ -350,15 +320,17 @@ def move_to_position(ids_positions):
 
         if dxl_comm_result != COMM_SUCCESS:
 
-            print(f"Error during SyncRead: {packetHandler.getTxRxResult(dxl_comm_result)}")
-
-
+            print(
+                f"Error during SyncRead: {packetHandler.getTxRxResult(dxl_comm_result)}"
+            )
 
         all_reached = True
 
         for dxl_id in ids_positions.keys():
 
-            dxl_present_position = groupSyncRead.getData(dxl_id, ADDR_PRESENT_POSITION, LEN_PRESENT_POSITION)
+            dxl_present_position = groupSyncRead.getData(
+                dxl_id, ADDR_PRESENT_POSITION, LEN_PRESENT_POSITION
+            )
 
             if dxl_present_position is None:
 
@@ -380,30 +352,22 @@ def move_to_position(ids_positions):
 
                 continue
 
-
-
             if abs(goal_position - dxl_present_position) > DXL_MOVING_STATUS_THRESHOLD:
 
                 all_reached = False
 
                 break
 
-
-
         if all_reached:
 
             break
 
+
 var = True
 while True:
 
-    
-
-
-    for _ in  range(5) :
+    for _ in range(5):
         # Move group 1 to max position
-
-        
 
         move_to_position({1: 1, 7: 1, 9: 1, 3: 0, 5: 0, 11: 0})
         move_to_position({2: 0, 8: 0, 10: 0, 4: 1, 6: 1, 12: 1})
@@ -415,42 +379,46 @@ while True:
         var = True
     if var:
         for _ in range(3):
-                    move_to_position({1: 0, 7: 1, 9: 0})
-                    move_to_position({2: 0, 8: 0, 10: 0})
-                    move_to_position({4: 1, 6: 1, 12: 1})
-                    move_to_position({1: 1, 7: 0, 9: 1})
-                    move_to_position({3: 1, 5: 0, 11: 1})
-                    move_to_position({4: 0, 6: 0, 12: 0})
-                    move_to_position({2: 1, 8: 1, 10: 1})
-                    move_to_position({3: 0, 5: 1, 11: 0})
+            move_to_position({1: 0, 7: 1, 9: 0})
+            move_to_position({2: 0, 8: 0, 10: 0})
+            move_to_position({4: 1, 6: 1, 12: 1})
+            move_to_position({1: 1, 7: 0, 9: 1})
+            move_to_position({3: 1, 5: 0, 11: 1})
+            move_to_position({4: 0, 6: 0, 12: 0})
+            move_to_position({2: 1, 8: 1, 10: 1})
+            move_to_position({3: 0, 5: 1, 11: 0})
     if not var:
         for _ in range(3):
-                move_to_position({1: 1, 7: 0, 9: 1})
-                move_to_position({2: 0, 8: 0, 10: 0})
-                move_to_position({4: 1, 6: 1, 12: 1})
-                move_to_position({1: 0, 7: 1, 9: 0})
-                move_to_position({3: 0, 5: 1, 11: 0})
-                move_to_position({4: 0, 6: 0, 12: 0})
-                move_to_position({2: 1, 8: 1, 10: 1})
-                move_to_position({3: 0, 5: 1, 11: 0})
+            move_to_position({1: 1, 7: 0, 9: 1})
+            move_to_position({2: 0, 8: 0, 10: 0})
+            move_to_position({4: 1, 6: 1, 12: 1})
+            move_to_position({1: 0, 7: 1, 9: 0})
+            move_to_position({3: 0, 5: 1, 11: 0})
+            move_to_position({4: 0, 6: 0, 12: 0})
+            move_to_position({2: 1, 8: 1, 10: 1})
+            move_to_position({3: 0, 5: 1, 11: 0})
 
 # Disable Dynamixel Torque
 
 for dxl_id in LIFT_SERVOS + ROTATE_SERVOS:
 
-    dxl_comm_result, dxl_error = packetHandler.write1ByteTxRx(portHandler, dxl_id, ADDR_TORQUE_ENABLE, TORQUE_DISABLE)
+    dxl_comm_result, dxl_error = packetHandler.write1ByteTxRx(
+        portHandler, dxl_id, ADDR_TORQUE_ENABLE, TORQUE_DISABLE
+    )
 
     if dxl_comm_result != COMM_SUCCESS:
 
-        print(f"Error disabling torque for Dynamixel {dxl_id}: {packetHandler.getTxRxResult(dxl_comm_result)}")
+        print(
+            f"Error disabling torque for Dynamixel {dxl_id}: {packetHandler.getTxRxResult(dxl_comm_result)}"
+        )
 
     elif dxl_error != 0:
 
-        print(f"Error disabling torque for Dynamixel {dxl_id}: {packetHandler.getRxPacketError(dxl_error)}")
-
+        print(
+            f"Error disabling torque for Dynamixel {dxl_id}: {packetHandler.getRxPacketError(dxl_error)}"
+        )
 
 
 # Close port
 
 portHandler.closePort()
-
